Backend/pyproject.toml
[project]
name = "meetings-to-document-backend"
version = "0.1.0"
description = "AI-powered application that transforms meeting recordings into structured Zuora Solution Design Documents"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "google-auth>=2.27.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.116.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "google-generativeai>=0.3.2",
    "assemblyai>=0.21.0",
    "crewai>=0.19.0",
    "langchain>=0.1.5",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.1",
    "celery[redis]>=5.3.6",
    "redis>=5.0.1",
    "httpx>=0.26.0",
    "aiofiles>=23.2.1",
    "structlog>=24.1.0",
    "python-dateutil>=2.8.2",
    "email-validator>=2.3.0",
    "greenlet>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "httpx>=0.26.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "workers", "document_schemas"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pyright>=1.1.407",
]
